Add a configuration to specify a filter file for what to deploy
Adds a -df -deployFilterfile configuration that is a file path to load
for telling the MavenRepositoryDeployer which GAVs to deploy. This file
contains a groupId:artifactId:version:ext (patterns supported) per
line, and is loaded into memory as a Set<Gav>. It implements the
Object.equals method on Gav and matcher support, so that we can compare
GAVs in order to know if it is in the filter file or not. If the
filter file is not specified, or doesnt exist, we do not filter at all.
The matching logic is made to be multi-threaded, in case the filter
file is large.

References: https://github.com/simpligility/maven-repository-tools/issues/81

Signed-off-by: Samuel Dacanay <sam.dacanay@chainguard.dev>